+++ cms/simulator/__init__.py
@@ -1,2 +1,4 @@
-from .simulator import ModelContext, Simulator, SimulatorError, computeExtraColumns
+from .base import ModelContext, SimulatorError
+from .simulator import Simulator
+from .observables import computeExtraColumns
 from .latex import modelExtendedLatex
