+++ cms/simulator/latex.py
@@ -38,6 +38,12 @@ def modelExtendedLatex(model: Model):
         out["compartments"][idx]["nameLatex"] = latex(Symbol(compartment.latex or compartment.name))
         out["compartments"][idx]["initLatex"] = latex(Symbol(normZero(compartment.latex or compartment.name)))
 
+    ## update observables latex
+    if (simulator.expressions):
+        for idx, (name, observable) in enumerate(simulator.observables.items()):
+            out["observables"][idx]["nameLatex"] = "%s" % latex(Symbol(name),symbol_names=symbols)
+            out["observables"][idx]["valueLatex"] = latex(observable,symbol_names =symbols)
+
     ## update expressions latex
     if (simulator.expressions):
         for idx, (name, expression) in enumerate(simulator.expressions.items()):
