+++ cms/simulator/simulator.py
@@ -3,62 +3,13 @@ from sympy import (
     sympify,
     Symbol,
     true as BTrue,
-    false as BFalse
+    false as BFalse,
+    Float as FloatT
 )
-from sympy.parsing.sympy_parser import parse_expr
-from sympy.core.numbers import Float as FloatT
-from cms.schemas import Model, Simulation
+from cms.schemas import Simulation
 from scipy.integrate import odeint
 import numpy as np
-
-class SimulatorError(Exception):
-    pass
-
-class SimulationResult:
-    def __init__(self, compartments, timeline, frames, param = None, paramValues = []):
-        self.compartments = compartments
-        self.timeline = timeline
-        self.frames = frames
-        self.param = param
-        self.paramValues = paramValues
-
-    @property
-    def isIterated(self):
-        return self.param is not None
-
-class ModelContext:
-    def __init__(self, model: Model):
-        self.params = {p.name: Symbol(p.name) for p in model.params}
-        
-        # initialize expression environment variables
-        self.expressionEnv = self.params.copy()
-        compartmentsInit = {f"{c.name}_0": Symbol(f"{c.name}_0") for c in model.compartments}
-        self.expressionEnv.update(compartmentsInit)
-        expressionsVars = {e.name: Symbol(e.name) for e in model.expressions}
-        self.expressionEnv.update(expressionsVars)
-
-        # TODO: Check recursion of expressions
-        self.expressions = {e.name: sympify(e.value, self.expressionEnv) for e in model.expressions}
-        self.compartments = {c.name: Symbol(c.name) for c in model.compartments}
-        self.preconditions = {p.predicate: sympify(p.predicate, self.expressionEnv) for p in model.preconditions}
-
-        # initialize reaction environment variables
-        self.reactionEnv = self.expressionEnv.copy()
-        self.reactionEnv.update(self.compartments)
-        self.reactionEnv["t"] = Symbol("t")
-        self.observables = {o.name: sympify(o.value, self.reactionEnv) for o in model.observables}
-
-        self.__initializeFormulas(model.reactions)
-
-    def __initializeFormulas(self, reactions):
-        self.formulas = {x: 0 for x in self.compartments}
-        self.odeVariables = set({})
-        for reaction in reactions:
-            funcExpr = sympify(reaction.function, self.reactionEnv)
-            self.odeVariables = self.odeVariables.union(funcExpr.free_symbols)
-            self.formulas[reaction.sfrom] -= funcExpr
-            self.formulas[reaction.sto] += funcExpr
-        self.odeVariables = tuple(self.odeVariables.difference(set(self.compartments.values())))
+from .base import ModelContext, SimulationResult, SimulatorError
 
 class Simulator:
     def __init__(self, context: ModelContext):
@@ -150,27 +101,3 @@ class Simulator:
         modelstr += "    return dz"
         modelcode = compile(modelstr, f"<odeModel>", "exec")
         return FunctionType(modelcode.co_consts[0], globals(), "ode_model")
-
-
-def __buildFunctionNP(compartments, name, expr):
-    # todo: support other environment variables. Only compartments are allowed in this version.
-    fstr = "def observable(row):\n"
-    fstr += f"    {', '.join(str(s) for s in compartments)}, *others = row\n"
-    fstr += f"    return {expr}"
-    fstrcode = compile(fstr, f"<{name}>", "exec")
-    return FunctionType(fstrcode.co_consts[0], globals(), name)
-
-def computeExtraColumns(context: ModelContext, result: SimulationResult):
-    if len(context.observables) > 0:
-        functions = [__buildFunctionNP(result.compartments, n, o) for n,o in context.observables.items()]
-        obsSize = len(context.observables)
-        compSize = len(result.compartments)
-        for findex, frame in enumerate(result.frames):
-            shape = list(frame.shape)
-            shape[1] += obsSize
-            newFrame = np.zeros(tuple(shape))
-            newFrame[:,:-obsSize] = frame
-            for idx, function in enumerate(functions):
-                newFrame[:, compSize + idx] = np.apply_along_axis(function, 1, newFrame)
-            result.frames[findex] = newFrame
-        result.compartments += list(context.observables.keys())
