+++ src/components/CurrentModel.vue
@@ -6,6 +6,7 @@
     <div v-if="currentModel.expressions">
       <h4>Where</h4>
       <div v-katex:display="expressions"></div>
+      <div v-katex:display="observables"></div>
     </div>
   </div>
 </template>
@@ -29,6 +30,17 @@ export default {
       if (this.currentModel.expressions) {
         out += "\\begin{aligned}";
         this.currentModel.expressions.forEach(expr => {
+          out += `${expr.nameLatex} = & ${expr.valueLatex}\\\\`;
+        });
+        out += "\\end{aligned}";
+      }
+      return out;
+    },
+    observables: function() {
+      let out = "";
+      if (this.currentModel.observables) {
+        out += "\\begin{aligned}";
+        this.currentModel.observables.forEach(expr => {
           out += `${expr.nameLatex} = & ${expr.valueLatex}\\\\`;
         });
         out += "\\end{aligned}";
